Possible Updates
The code looks good and is working as intended. However, there are some things that could be improved or refactored:

The getCoordinates method could be split up into smaller methods to make it more modular and easier to read.
The error message could be shown in a better way, like in a modal or as a toast notification.
The form could be given a more descriptive aria-label attribute for accessibility.
The process.env.REACT_APP_LOCATIONIQ_API_KEY could be stored in a separate file and imported to avoid having it hard-coded in the component.
The conditional rendering of the location information could be moved to a separate component for better reusability.